feat: 컨트랙트 함수 추가
