+++ src/router/Router.jsx
@@ -10,6 +10,7 @@ import Collection from '../pages/collection/Collection';
 import NFTDetails from '../pages/detail/NFTDetails';
 import Account from '../pages/account/Account';
 import BuyNFTDetails from '../components/explore/BuyNFTDetails';
+import MyNFTDetails from '../components/mypage/MyNFTDetails';
 
 export default function Router({ web3, collapsed, setCollapsed, account }) {
   return (
@@ -28,6 +29,10 @@ export default function Router({ web3, collapsed, setCollapsed, account }) {
       <Route path="/assets/art" element={<Explore />} />
       <Route path="/assets/collectibles" element={<Explore />} />
       <Route
+        path="/assets/buy/ethereum/:asset_contract/:token_id"
+        element={<BuyNFTDetails web3={web3} setCollapsed={setCollapsed} account={account} />}
+      />
+      <Route
         path="/assets/create"
         element={<Create web3={web3} setCollapsed={setCollapsed} account={account} />}
       />
@@ -35,14 +40,14 @@ export default function Router({ web3, collapsed, setCollapsed, account }) {
         path="/mypage"
         element={<Account web3={web3} setCollapsed={setCollapsed} account={account} />}
       />
+      <Route
+        path="/assets/mypage/ethereum/:asset_contract/:token_id"
+        element={<MyNFTDetails web3={web3} setCollapsed={setCollapsed} account={account} />}
+      />
       <Route path="/collection/:collection_key" element={<Collection />} />
       <Route
         path="/assets/ethereum/:asset_contract/:token_id"
         element={<NFTDetails web3={web3} setCollapsed={setCollapsed} account={account} />}
-      />
-      <Route
-        path="/assets/buy/ethereum/:asset_contract/:token_id"
-        element={<BuyNFTDetails web3={web3} setCollapsed={setCollapsed} account={account} />}
       />
     </Routes>
   );
